requirements.txt
gunicorn==20.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
Flask==3.0.3
future==1.0.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huepy==1.2.1
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mock==5.1.0
moviepy==1.0.3
numpy==2.0.0
oauthlib==3.2.2
openai==1.33.0
outcome==1.3.0.post0
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
priority==2.0.0
proglog==0.1.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
PySocks==1.7.1
pytest==8.2.2
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
Quart==0.19.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.3
rsa==4.9
schedule==1.2.2
selenium==4.21.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.1
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
